before sync with hub

+++ app/controllers/application_controller.rb
@@ -38,6 +38,16 @@ class ApplicationController < ActionController::Base
     end
   end
 
+  # current_user must be part of this meet as a pending user
+  def pending_meet_member(id=nil)
+    @meet ||= Meet.find_by_id(id||params[:id])
+    if (!@meet || (!@meet.include_pending_user?(current_user) && !admin_user?))
+      render_unauthorized
+    else
+      attach_meet_mview(current_user, @meet)
+    end
+  end
+
   # current_user must be own this mpost
   def authorized_mpost_owner(id=nil)
     @mpost ||= Mpost.find_by_id(id||params[:id])
@@ -53,8 +63,6 @@ class ApplicationController < ActionController::Base
     @chatter ||= Chatter.find_by_id(id||params[:id])
     if (!@chatter || (!current_user_id?(@chatter.user_id) && !admin_user?))
       render_unauthorized
-    elsif (!@chatter.active? && !admin_user?)
-      render_removed
     end
   end
 
@@ -70,15 +78,13 @@ class ApplicationController < ActionController::Base
 
   # current_user must be part of chatter's meet and this chatter must be either
   # a topic or a comment to a topic. Can not add a comment to another comment.
-  def authroized_chatter
+  def authorized_chatter
     if params[:meet_id].present?
       authorized_meet_member(params[:meet_id])
     else
       @topic = Chatter.find_by_id(params[:chatter_id])
       if (!@topic || !@topic.topic?)
         render_unauthorized
-      elsif (!@topic.active? && !admin_user?)
-        render_removed
       else
         authorized_meet_member(@topic.meet_id)
       end
@@ -111,76 +117,92 @@ class ApplicationController < ActionController::Base
   class InvalidAssert < Exception; end
   class UnauthorizedAssert < Exception; end
   class InternalErrorAssert < Exception; end
+  class RemoveAssert < Exception; end
   rescue_from InvalidAssert do |exception| render_invalid; end
   rescue_from UnauthorizedAssert do |exception| render_unauthorized; end
   rescue_from InternalErrorAssert do |exception| render_internal_error; end
-  def assert_invalid(val=true, &block)
-    raise InvalidAssert if (!val || (block && !block.call))
+  rescue_from RemoveAssert do |exception| render_remove; end
+  def assert_invalid(val=true, options={}, &block)
+    render_invalid({:raise=>InvalidAssert}.merge(options)) if (!val || (block && !block.call))
   end
-  def assert_unauthorized(val=true, &block)
-    raise UnauthorizedAssert if (!val || (block && !block.call))
+  def assert_unauthorized(val=true, options={}, &block)
+    render_unauthorized({:raise=>UnauthorizedAssert}.merge(options)) if (!val || (block && !block.call))
   end
-  def assert_internal_error(val=true, &block)
-    raise InternalErrorAssert if (!val || (block && !block.call))
+  def assert_internal_error(val=true, options={}, &block)
+    render_internal_error({:raise=>InternalErrorAssert}.merge(options)) if (!val || (block && !block.call))
+  end
+  def assert_remove(val=true, options={}, &block)
+    render_remove({:raise=>RemoveAssert}.merge(options)) if (!val || (block && !block.call))
   end
 
-  def filter_params(:context, options = {})
-    @filtered_params = params[:context].nil ? params : params[:context]
-    strips = options[:strip]
-    if strips.present?
-      strips.each {|key|
-        @filtered_params[key].strip! if @filtered_params[key].present?
-      }
-    end
+  def filter_params(context, options = {})
+    @filtered_params = params[context].nil? ? params.clone : params[context].clone
+    @filtered_params.delete(:controller)
+    @filtered_params.delete(:action)
+    options[:strip].each {|key|
+      @filtered_params[key].strip! if @filtered_params[key].present?
+    } if options[:strip].present?
   end
   
-  def delete_mposts(user, mposts)
+  def delete_mposts(mposts)
     return unless mposts.present?
     mposts.each {|mpost|
-      #user.mposts.destroy(mpost); @meet.mposts.destroy(mpost); mpost.destroy
+      #next if mpost.destroyed?
+      #user, meet = mpost.user, mpost.meet
+      #user.mposts.destroy(mpost)
+      #meet.mposts.destroy(mpost) if meet.present?
+      #mpost.destroy
       mpost.delete; mpost.save
     }
   end
-  def delete_mviews(user, mviews)
+  def delete_mviews(mviews)
     return unless mviews.present?
     mviews.each {|mview|
+      next if mview.destroyed?
+      user, meet = mview.user, mview.meet
       user.mviews.destroy(mview)
       meet.mviews.destroy(mview)
       mview.destroy
     }
   end
-  def delete_invitations(user, invitations)
+  def delete_invitations(invitations)
     return unless invitations.present?
     invitations.each {|invitation|
+      next if invitation.destroyed?
+      user, meet = invitation.user, invitation.meet
       user.invitations.destroy(invitation)
-      meet.invitations.destroy(invitation)
+      meet.invitations.destroy(invitation) if meet.present?
       invitation.destroy
     }
   end
-  def delete_chatters(user, chatters)
+  def delete_chatters(chatters)
     update_meets = Set.new
     update_topics = Set.new
-    invitations.each {|invitation|
-      next if invitation.destroyed?
+    chatters.each {|chatter|
+      next if chatter.destroyed?
+      user, meet = chatter.user, chatter.meet
       topic = chatter.topic
-      if topic.present?
+      # Be careful, a topic can be an orphan (no user).
+      if (user.present? && chatter.topic?)
         # Delete all current user's comments under this topic first
         chatter.comments.to_a.each {|comment|
-          if comment.user_id == current_user.id
+          if comment.user_id == user.id
             meet.chatters.destroy(comment)
             meet.chatters.destroy(comment)
-            topic.chatters.destroy(comment)
+            chatter.comments.destroy(comment)
             comment.destroy
           end
         }
       end
-      if (!topic || chatter.comments_count == 0)
+      if (!chatter.topic? || chatter.comments.empty?)
         # Delete the chatter if it is not a topic or it is has no comments
         meet.chatters.destroy(chatter)
         update_meets << meet
-        user.chatters.destroy(chatter)
-        topic.comments.destroy(chatter) if topic.present?
-        update_topics << topic
+        user.chatters.destroy(chatter) if user.present?
+        if topic.present?
+          topic.comments.destroy(chatter)
+          update_topics << topic
+        end
         chatter.destroy
       else
         # Can not delete it, otherwise it will wipe out all comments under it.
@@ -189,7 +211,7 @@ class ApplicationController < ActionController::Base
         chatter.save
       end
       update_meets.each {|meet|
-        @meet.opt_lock_protected {
+        meet.opt_lock_protected {
           meet.update_chatters_count
           meet.save
         }
@@ -199,22 +221,31 @@ class ApplicationController < ActionController::Base
         topic.update_comments_count
         topic.save
       }
+    }
   end
 
   private
 
     def render_error(file, status, options={})
-      except, only = options[:except], options[:only]
+      except, only, raise_exception = options[:except], options[:only], options[:raise]
       except = [except] if (except && !except.is_a?(Array))
       only = [only] if (only && !only.is_a?(Array))
       is_html = (!except || !except.include?(:html)) &&
                 (!only   || only.include?(:html))
       is_json = (!except || !except.include?(:json)) &&
                 (!only   || only.include?(:json))
-      respond_to do |format|
-        format.html { redirect_to file } if is_html
-        #format.html { render :file=>file, :status=>status } if is_html
-        format.json { head status } if is_json
+      format = params[:format] || request.format
+      format = format.to_s
+      format = format.slice((format.rindex('/')||-1)+1..-1);
+      if raise_exception.present?
+        raise raise_exception if (is_html && format == "html")
+        raise raise_exception if (is_json && format == "json")
+      else
+        respond_to do |format|
+          #format.html { render :file=>file, :status=>status; raise AssertException } if is_html
+          format.html { redirect_to file } if is_html
+          format.json { head status } if is_json
+        end
       end
     end
